Properly initialize fugitive in existing buffers
Fugitive installs buffer-local commands (Gstatus, Gblame, etc.)
when it has detected that a buffer is in a git repository.  The problem
with lazy-initialization is that if you open two buffers (both in git
repositories), call :Gstatus in one, and then try to run :Gstatus in
the other you'll simply get an error about "Not an editor command"

This adds a hook_post_source to call the fugitive git detection
in all existing buffers after loading fugitive.