verify yaml2json before using it
This addresses #44 which had an issue when yaml2json in the PATH was not
something as simple as reads from stdin/writes to stdout.  In order to
test, let's just put together a really simple yaml file consisting of:

    ---
    test: 1

and make sure we can parse it and json_decode it.